Revert "refactor(gerichtskosten): move gerichtskosten.ts to shared folder"
This reverts commit cac9075750764a278c7bf27cd25ab889081fcda2.
